extract_coordinate_values {crownsegmentr}	R Documentation

Extract coordinate data from an object

Description

This function extracts three numeric columns from the input table. If possible, columns which are named x/X, y/Y, or z/Z.

Usage

extract_coordinate_values(coordinate_table)

Arguments

coordinate_table
An object which is valid according to validate_coordinate_table (i.e. data.frame-like and contains at least three numeric columns).

Value

A data.frame with just three columns that are expected to hold the x-, y-, and z-coordinates in that order.

[Package crownsegmentr version 0.4.4 Index]