validate_write_crown_id_also_to_file_for_LAScatalogs {crownsegmentr}	R Documentation

Ensures that crown IDs are written to output files of the LAScatalog

Description

Issues a warning if the user wanted to write the output to files but not store IDs of segmented bodies.

Usage

validate_write_crown_id_also_to_file_for_LAScatalogs(
  write_crown_id_also_to_file,
  LAScatalog
)

Arguments

write_crown_id_also_to_file
The to-be-validated parameter.

LAScatalog
The LAScatalog-class object whose settings are compared to the value of write_crown_id_also_to_file.

Value

A possibly corrected value for write_crown_id_also_to_file.

[Package crownsegmentr version 0.4.4 Index]