collect_scale_n_offset_of_LAScatalog_files {crownsegmentr}	R Documentation

Get the scale and offset values of all files referenced by a LAScatalog

Description

Get the scale and offset values of all files referenced by a LAScatalog

Usage

collect_scale_n_offset_of_LAScatalog_files(LAScatalog)

Value

A data.table with 7 columns. The first three columns hold the x, y, and z scale factors, the next three columns hold the x, y, and z offsets and the last column holds the file paths. There is one row for each referenced file.

[Package crownsegmentr version 0.4.4 Index]